Refactor shared types and gravatar utils, add avatar URL validation
- Create src/types/index.ts with shared AccessLevel, ProfileStatus, MemberProfile, TeamMemberProfile
- Extract gravatar helpers into src/utils/gravatar.ts shared util
- Simplify safeAvatarUrl to only allow https/http (no local path exception)
- Add Gravatar URL validation to GeneralInfoFormModal on avatar save
- Restore package.json and package-lock.json to match dev branch
