+++ src/components/CompleteProfile/CompleteProfile.tsx
@@ -11,6 +11,7 @@ import { XTwitter } from '../Icons/Social/XTwitter.tsx';
 import StepBar from '../StepBar/StepBar.tsx';
 import './CompleteProfile.css';
 import { safeAvatarUrl } from '../../utils/safeAvatarUrl.ts';
+import { getGravatarAvatarUrl, getGravatarUrlType } from '../../utils/gravatar.ts';
 
 interface SocialIcons {
 	id: string;
@@ -36,30 +37,7 @@ type UpdateProfileParams = Omit<ProfileParams, 'email'>;
 
 const platformEnum = ['site', 'slack', 'linkedin', 'github', 'portfolio', 'codepen', 'instagram', 'threads', 'facebook', 'bluesky', 'mastodon', 'twitter', 'devto'];
 
-const isGravatarHost = (hostname: string): boolean => {
-	const normalizedHostname = hostname.toLowerCase();
-	return normalizedHostname === 'gravatar.com' || normalizedHostname.endsWith('.gravatar.com');
-};
 
-const getGravatarUrlType = (value?: string): 'empty' | 'image' | 'invalid' | 'profile' => {
-	if (!value?.trim()) { return 'empty'; }
-
-	try {
-		const url = new URL(value.trim());
-		const [firstPath = '', secondPath, thirdPath] = url.pathname.split('/').filter(Boolean);
-
-		if (url.protocol !== 'https:' || !isGravatarHost(url.hostname)) { return 'invalid'; }
-		if (firstPath === 'avatar' && secondPath) { return 'image'; }
-		if ((url.hostname === 'gravatar.com' || url.hostname === 'www.gravatar.com') && firstPath && !secondPath) { return 'profile'; }
-		if (url.hostname === 'api.gravatar.com' && firstPath === 'v3' && secondPath === 'profiles' && thirdPath) { return 'profile'; }
-
-		return 'invalid';
-	} catch {
-		return 'invalid';
-	}
-};
-
-// Validation helpers
 /**
  * Checks whether a birthday string is in the valid MM-DD format.
  *
@@ -121,43 +99,6 @@ const updateProfile = async (data: UpdateProfileParams, profileId: string) => {
 	}
 };
 
-const getGravatarProfileSlug = (value: string): string | null => {
-	try {
-		const url = new URL(value.trim());
-		const [firstPath = '', secondPath = '', thirdPath = '', extraPath] = url.pathname.split('/').filter(Boolean);
-
-		if (url.protocol !== 'https:') {
-			return null;
-		}
-
-		if (['gravatar.com', 'www.gravatar.com'].includes(url.hostname.toLowerCase()) && firstPath !== 'avatar' && !secondPath) {
-			return firstPath.replace(/\.card$/iu, '');
-		}
-
-		if (url.hostname === 'api.gravatar.com' && firstPath === 'v3' && secondPath === 'profiles' && thirdPath && !extraPath) {
-			return thirdPath;
-		}
-
-		return null;
-	} catch {
-		return null;
-	}
-};
-
-const getGravatarAvatarUrl = async (profileUrl: string): Promise<string | null> => {
-	const slug = getGravatarProfileSlug(profileUrl);
-	if (!slug) { return null; }
-
-	try {
-		const response = await fetch(`https://api.gravatar.com/v3/profiles/${encodeURIComponent(slug)}`);
-		if (!response.ok) { return null; }
-
-		const profile = await response.json() as { avatar_url?: string };
-		return profile.avatar_url ?? null;
-	} catch {
-		return null;
-	}
-};
 
 // eslint-disable-next-line max-lines-per-function
 const CompleteProfile = () => {
