+++ src/components/ProfileCard/ProfileCard.tsx
@@ -3,17 +3,7 @@ import './ProfileCard.css';
 import { Facebook } from '../Icons/Social/Facebook.tsx';
 import { LinkedIn } from '../Icons/Social/LinkedIn.tsx';
 import { XTwitter } from '../Icons/Social/XTwitter.tsx';
-
-export interface MemberProfile {
-	id: string;
-	name: string;
-	email: string;
-	avatar?: string;
-	socialLinks?: Record<string, string>;
-	description?: string;
-	happenedAt: string;
-	insertedAt: string;
-}
+import type { MemberProfile } from '../../types/index.ts';
 
 // Maps user's social media to a React Component containing the icon for the respective social media
 // => It picks up an icon based on the user's social media handle in database
